+++ MatrixPinMem.cu
@@ -5,7 +5,7 @@ Matrix::Matrix (int rows, int cols, unsigned int flag, int supply) :
   if (rows == 0 || cols == 0) {
     CHECK_NULL(NULL);
   }
-  CHECK_CUDA(cudaHostAlloc((void**)&e, m * cols * sizeof(double), flag));
+  CHECK_CUDA(cudaHostAlloc((void**)&e, m * cols * sizeof(float), flag));
   for (int j = 0; j < cols; j++) {
     for (int i = 0; i < rows; i++) {
       e[i + j * m] = 0.0;
@@ -16,7 +16,7 @@ Matrix::Matrix (int rows, int cols, unsigned int flag, int supply) :
 Matrix::Matrix(Matrix const &input, unsigned int flag, int supply) :
     rows(input.rows), cols(input.cols), supply(supply), m(input.m + supply) {
   const uint full = m * cols;
-  CHECK_CUDA(cudaHostAlloc((void**)&e, full * sizeof(double), flag));
+  CHECK_CUDA(cudaHostAlloc((void**)&e, full * sizeof(float), flag));
 
   for (int j = 0; j < cols; j ++) {
     for (int i = 0; i < rows; i++) {
